Fixed problem when copying CSS resources, fixes #6902

+++ org.eclipse.virgo.ide.runtime.core/src/org/eclipse/virgo/ide/runtime/core/ches/VirgoToolingHook.java
@@ -648,6 +648,10 @@ public class VirgoToolingHook {
                 File source = child.getLocation().toFile();
                 List<File> targets = lookup(source);
                 for (File target : targets) {
+                    if (!source.exists()) {
+                        continue; // make sure the resource has not been deleted meanwhile, #6902
+                    }
+
                     Files.copy(source.toPath(), target.toPath(), StandardCopyOption.REPLACE_EXISTING);
                     logInfo("Copying " + source.getName() + " --> " + target.getAbsolutePath());
                 }
